requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
pydantic-settings>=2.0.0
sqlmodel>=0.0.18
telethon>=1.34.0
chromadb>=0.5.0
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-ollama>=0.1.0
llama-index-llms-ollama>=0.1.0
langdetect>=1.0.9
apscheduler>=3.10.0
aiosmtplib>=3.0.0
weasyprint>=62.0
jinja2>=3.1.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
websockets>=12.0
aiofiles>=23.0.0
